+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - score belongs to performed_segments ?
 - having scores for skater datatable
 - components feature spec
+- update competition: season from/to
 
 ### new feature